backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
sqlalchemy==2.0.25
sentence-transformers==2.3.1
scikit-learn==1.4.0
numpy==1.26.3
python-dotenv==1.0.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
